refactor to align with Milo
