+++ express/features/direct-path-to-product/direct-path-to-product.css
@@ -1,0 +1,179 @@
+.bmtp-container {
+    position: absolute;
+    right: 8px;
+    top: 72px;
+    background-color: var(--color-white);
+    width: 376px;
+    border: 1px solid #e8e8e8;
+    border-radius: 10px;
+    z-index: 99;
+    overflow: hidden;
+}
+
+.bmtp-container .bmtp-header {
+    display: flex;
+    align-items: center;
+    font-weight: 700;
+    padding: 16px;
+}
+
+.bmtp-container .bmtp-header-icon {
+    margin-right: 12px;
+}
+
+.bmtp-container .bmtp-header-icon .icon {
+    height: 32px;
+    width: 32px;
+    display: block;
+}
+
+.bmtp-container .bmtp-progress-bg {
+    height: 4px;
+    background-color: #d6d6d6;
+}
+
+.bmtp-container .bmtp-progress-bg .bmtp-progress-bar {
+    background-color: var(--color-info-accent);
+    width: 100%;
+    height: 100%;
+    -webkit-animation: loginRedirectProgress 2s ease;
+    -moz-animation: loginRedirectProgress 2s ease;
+    animation: loginRedirectProgress 2s ease;
+}
+
+.bmtp-container .profile-wrapper {
+    display: flex;
+    align-items: center;
+    padding: 16px;
+}
+
+.bmtp-container .profile-wrapper .profile-img-container {
+    position: relative;
+    margin-right: 14px;
+}
+
+.bmtp-container .profile-wrapper .profile-img-container::before {
+    content: "";
+    animation: rotateCircle 1.5s linear infinite;
+    animation-play-state: inherit;
+    border: solid 2px var(--color-gray-200);
+    border-bottom-color: #33AB84;
+    border-radius: 50%;
+    height: 30px;
+    width: 30px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate3d(-50%, -50%, 0);
+    will-change: transform;
+}
+
+.bmtp-container.done .profile-wrapper .profile-img-container::before {
+    border: none;
+    height: 32px;
+    width: 32px;
+    animation: unset;
+}
+
+.bmtp-container.done .profile-wrapper .profile-img-container::after {
+    background-image: url('/express/icons/success-check.svg');
+    content: "";
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate3d(-50%, -50%, 0);
+    height: 32px;
+    width: 32px;
+    border-radius: 50%;
+}
+
+.bmtp-container .profile-wrapper .profile-img-container img {
+    display: block;
+    height: 30px;
+    width: 30px;
+    border-radius: 28px;
+}
+
+.bmtp-container .profile-text-wrapper {
+    font-size: var(--body-font-size-s);
+}
+
+.bmtp-container .profile-text-wrapper .profile-name,
+.bmtp-container .profile-text-wrapper .profile-email {
+    display: block;
+}
+
+.bmtp-container .notice-wrapper {
+    padding: 16px;
+    font-size: var(--body-font-size-s);
+    background-color: var(--color-gray-100);
+    display: flex;
+    justify-content: space-between;
+}
+
+.bmtp-container .notice-btn {
+    font-weight: 700;
+    color: var(--color-info-accent);
+    cursor: pointer;
+}
+
+.bmtp-container .notice-btn:hover {
+    color: var(--color-info-accent-hover);
+}
+
+@-webkit-keyframes loginRedirectProgress {
+    0% {
+        width: 0;
+    }
+
+    100% {
+        width: 100%
+    }
+}
+
+@-moz-keyframes loginRedirectProgress {
+    0% {
+        width: 0;
+    }
+
+    100% {
+        width: 100%
+    }
+}
+
+@keyframes loginRedirectProgress {
+    0% {
+        width: 0;
+    }
+
+    100% {
+        width: 100%
+    }
+}
+
+@-webkit-keyframes rotateCircle {
+    0% {
+        transform: translate3d(-50%, -50%, 0) rotate(0deg);
+    }
+    100% {
+        transform: translate3d(-50%, -50%, 0) rotate(360deg);
+    }
+}
+
+@-moz-keyframes rotateCircle {
+    0% {
+        transform: translate3d(-50%, -50%, 0) rotate(0deg);
+    }
+    100% {
+        transform: translate3d(-50%, -50%, 0) rotate(360deg);
+    }
+}
+
+@keyframes rotateCircle {
+    0% {
+        transform: translate3d(-50%, -50%, 0) rotate(0deg);
+    }
+    100% {
+        transform: translate3d(-50%, -50%, 0) rotate(360deg);
+    }
+}
