+++ express/scripts/delayed.js
@@ -10,12 +10,11 @@
  * governing permissions and limitations under the License.
  */
 
-import BlockMediator from './block-mediator.min.js';
 import {
   fetchPlaceholders,
-  getIconElement,
   createTag,
   getDevice,
+  getMetadata,
 } from './utils.js';
 
 export const loadExpressProduct = async () => {
@@ -27,100 +26,37 @@ export const loadExpressProduct = async () => {
   document.body.append(iframe);
 };
 
-async function loadLoginUserAutoRedirect() {
-  let followThrough = true;
+async function isLoggedInDelayed() {
+  const userProfile = window.adobeProfile?.getUserProfile();
   const placeholders = await fetchPlaceholders();
-  const buildRedirectAlert = async (profile) => {
-    const container = createTag('div', { class: 'bmtp-container' });
-    const headerWrapper = createTag('div', { class: 'bmtp-header' });
-    const headerIcon = createTag('div', { class: 'bmtp-header-icon' }, getIconElement('cc-express'));
-    const headerText = createTag('span', { class: 'bmtp-header-text' }, placeholders['bmtp-header']);
-    const progressBg = createTag('div', { class: 'bmtp-progress-bg' });
-    const progressBar = createTag('div', { class: 'bmtp-progress-bar' });
-    const profileWrapper = createTag('div', { class: 'profile-wrapper' });
-    const profilePhotoCont = createTag('div', { class: 'profile-img-container' });
-    const profilePhoto = createTag('img', { src: profile.avatar });
-    const profileTextWrapper = createTag('div', { class: 'profile-text-wrapper' });
-    const profileName = createTag('strong', { class: 'profile-name' }, profile.display_name);
-    const profileEmail = createTag('span', { class: 'profile-email' }, profile.email);
-    const noticeWrapper = createTag('div', { class: 'notice-wrapper' });
-    const noticeText = createTag('span', { class: 'notice-text' }, placeholders['bmtp-cancel-text']);
-    const noticeBtn = createTag('a', { class: 'notice-btn' }, placeholders.cancel);
+  const autoRedirect = ['yes', 'true', 'Y', 'on'].includes(getMetadata('direct-path-to-product'));
+  const autoRedirectLanguageFound = placeholders.cancel || placeholders['bmtp-header'] || placeholders['bmtp-cancel-text'];
+  const isDesktop = getDevice() !== 'desktop';
 
-    const photoInGNav = document.querySelector('header .Profile-thumbnail');
-    if (photoInGNav) {
-      [, profilePhoto.src] = photoInGNav.style.backgroundImage.match('url\\("([^"]+)"\\)');
-    }
-
-    headerWrapper.append(headerIcon, headerText);
-    progressBg.append(progressBar);
-    profilePhotoCont.append(profilePhoto);
-    profileWrapper.append(profilePhotoCont, profileTextWrapper);
-    profileTextWrapper.append(profileName, profileEmail);
-    noticeWrapper.append(noticeText, noticeBtn);
-    container.append(headerWrapper, progressBg, profileWrapper, noticeWrapper);
-
-    const header = document.querySelector('header');
-    header.append(container);
-
-    noticeBtn.addEventListener('click', () => {
-      container.remove();
-      followThrough = false;
-      localStorage.setItem('no-bmtp', '3');
-    });
-
-    return container;
-  };
-
-  const initRedirect = (container) => {
-    if (!followThrough) return;
-
-    container.classList.add('done');
-
-    const primaryCtaUrl = BlockMediator.get('primaryCtaUrl')
-      || document.querySelector('a.button.xlarge.same-as-floating-button-CTA, a.primaryCTA')?.href;
-
-    if (primaryCtaUrl) {
-      window.location.assign(primaryCtaUrl);
-    } else {
-      window.assign('https://new.express.adobe.com');
-    }
-  };
-
-  const profile = window.adobeProfile.getUserProfile();
-
-  const optOutCounter = localStorage.getItem('no-bmtp');
-  if (!optOutCounter || optOutCounter === '0') {
-    const container = await buildRedirectAlert(profile);
-    setTimeout(() => {
-      if (container) initRedirect(container);
-    }, 2000);
-  }
-
-  if (optOutCounter && optOutCounter !== '0') {
-    const counterNumber = parseInt(optOutCounter, 10);
-    localStorage.setItem('no-bmtp', (counterNumber - 1).toString());
-  }
-}
-
-export function loadDelayedLoggedIn(DELAY = 3000) {
-  return new Promise((resolve) => {
-    loadExpressProduct();
-    setTimeout(() => {
-      loadLoginUserAutoRedirect();
-      resolve();
-    }, DELAY);
-  });
+  return userProfile && autoRedirect && autoRedirectLanguageFound && isDesktop;
 }
 
 /**
  * Executes everything that happens a lot later, without impacting the user experience.
  */
 export default function loadDelayed(DELAY = 3000) {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      loadExpressProduct();
-      resolve();
-    }, DELAY);
+  isLoggedInDelayed().then(async (bringMeToProduct) => {
+    if (bringMeToProduct) {
+      const { default: loadLoginUserAutoRedirect } = await import('../features/direct-path-to-product/direct-path-to-product.js');
+      return new Promise((resolve) => {
+        loadExpressProduct();
+        setTimeout(() => {
+          loadLoginUserAutoRedirect();
+          resolve();
+        }, DELAY);
+      });
+    }
+
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        loadExpressProduct();
+        resolve();
+      }, DELAY);
+    });
   });
 }
